Unify QA tooling with koopmans2 (ruff + pre-commit)
Lift koopmans2's ruff config into aiida-koopmans2 so both repos now run
the same lint/format policy on every commit:

- line-length = 100 (was 120)
- Rule set now matches koopmans2: F, E, W, C90, I, UP, D, DOC (preview),
  B, S, T20, N, ERA, NPY, RUF, C4 — replaces the leaner F/E/W/I/N/PL/RUF set.
- DOC (pydoclint) is opted into via ``preview = true`` +
  ``explicit-preview-rules = true`` — preview mode on, but only the rule
  we want is activated. Without this, "Selection `DOC` has no effect"
  warnings appear but pydoclint does nothing.
- Per-file ignores: D101/D102/D103/D107/S101 on tests (descriptive names
  replace per-method docstrings, assert is the idiom), F401/F403 on
  package __init__ files, N802 on @task.graph CamelCase names in
  workgraphs/ (matching upstream Wannier90TaskViaBuilder etc.).
- ruff version pin bumped 0.4.3 → 0.14.8 in the hatch static-analysis env.

Pre-commit switched from ``hatch fmt`` delegation to direct
``astral-sh/ruff-pre-commit`` + universal hygiene hooks (trailing whitespace,
end-of-file, yaml/toml/merge-conflict checks, added-large-files guard).

Code fixups to satisfy the stricter rule set:
- ``_BLOCKED_KEYS`` annotated as ``ClassVar`` (RUF012)
- Dropped redundant ``int(round(...))`` cast in count_electrons (RUF046)
- Split ``parse`` into ``_emit_eigenvalues`` / ``_emit_lambdas`` helpers,
  and ``prepare_for_submission`` into ``_inject_owned_keys`` /
  ``_build_local_copy_list`` / ``_write_alpha_files`` /
  ``_build_remote_symlink_list`` / ``_build_retrieve_list`` helpers (C901).
- ``_parse_stdout`` kept as a single function with ``# noqa: C901`` — its
  branching is inherent to kcp.x's multi-pattern stdout.
- ``xml.etree.ElementTree.fromstring`` flagged as S314 silenced with
  ``# noqa`` + explanation (input is our own trusted kcp.x output).
- Docstring fixes (D401 imperative mood), commented-code cleanup (ERA001),
  long-line split (E501), dict-literal rewrite (C408).

+++ src/aiida_koopmans/utils.py
@@ -31,9 +31,7 @@ def resolve_pseudo_family(family_label: str, structure: orm.StructureData) -> di
     from aiida_pseudo.groups.family.pseudo import PseudoPotentialFamily
 
     family = (
-        orm.QueryBuilder()
-        .append(PseudoPotentialFamily, filters={"label": family_label})
-        .one()[0]
+        orm.QueryBuilder().append(PseudoPotentialFamily, filters={"label": family_label}).one()[0]
     )
     return family.get_pseudos(structure=structure)
 
@@ -59,7 +57,7 @@ def count_electrons(
     nelec_total = 0.0
     for site in structure.sites:
         nelec_total += float(pseudos[site.kind_name].z_valence)
-    nelec = int(round(nelec_total))
+    nelec = round(nelec_total)
     if nelec != nelec_total:
         raise ValueError(
             f"Non-integer total valence charge {nelec_total} from pseudos — "
@@ -70,9 +68,7 @@ def count_electrons(
 
     m = tot_magnetization if tot_magnetization is not None else 0
     if (nelec + m) % 2 or (nelec - m) % 2:
-        raise ValueError(
-            f"nelec={nelec}, tot_magnetization={m} give non-integer spin populations."
-        )
+        raise ValueError(f"nelec={nelec}, tot_magnetization={m} give non-integer spin populations.")
     return nelec, (nelec + m) // 2, (nelec - m) // 2
